requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
pydantic==2.9.0
pydantic-settings==2.6.0
python-dotenv==1.0.1
supabase==2.8.0
torch==2.8.0
torchvision==0.23.0
open-clip-torch==2.28.0
Pillow==11.0.0
geoalchemy2==0.14.2 
pgvector==0.3.0